Initial commit for User model unit test
Note that this is a work in progress. Some boilerplate code will likely be moved
to a trait later so that all the model unit tests share the same stuff.

The user model unit test is still incomplete but I got it working.

TODO: Figure out how to use the After and Before in Specs2. Doesn't seem to be
working right now.